Generate idl for the canisters according to the new canister api
With canister-as-trait approach for generating canister api we're
faced with a problem of generating idls for them. Currently, the
way it's done is that we have an internal lazy static variable
called `METHODS` that is filled up with method info each time the
update/query method is expanded and later this variable is queried
when `generate_idl!()` macro is called. The problem with it is that
this `METHODS` variable is crate-local, hence when `ic-helpers`
is imported, it has its own `METHODS` than the original canister's.
Also, since this `METHODS` variable cannot be exported from a
`proc-macro`-type crate (in this case `ic-canister-macros`) and
cannot be moved anywhere else, we need to come up with a solution
on how we can export the contents of that `METHODS` variable and
send it to the other canisters. Since macroses cannot return
values, the only approach that we can take is to generate some code
that returns the method info accordingly. Hence, we have an `Idl`
struct, that will be filled by the code, generated at compile time
by `generate_idl!()` function. This code will be a method of
`FactoryCanister` trait, meaning that this function will be
constant. When `ic-helpers` is imported, the code for `get_idl!()`
will be already compiled, which later can be used in other
canisters as expected.